refactor: consolidate forecast provider implementation
- Remove redundant _log_forecast_solar_rate_limit() in Forecast.Solar (duplicate logging)
- Add daily_kwh calculation to Open-Meteo provider for consistency across all providers
- Standardize return type: all providers now return Tuple[Dict[Dict], Dict[Dict]]
- Unify logging pattern: start + end logs with entry counts across all providers
- Translate all docstrings and comments to German
- Simplify store logic: remove _calculate_daily_kwh() (now done by providers)
- Consistent implementation pattern for future providers
