+++ packages/modules/common/store/_forecast.py
@@ -21,40 +21,6 @@ def _parse_forecast_timestamp(timestamp: str) -> Optional[datetime]:
     except (TypeError, ValueError):
         return None
 
-
-def _calculate_daily_kwh(values: Dict[str, float]) -> Dict[str, float]:
-    points: list[tuple[datetime, float]] = []
-    for timestamp, value in values.items():
-        parsed_timestamp = _parse_forecast_timestamp(timestamp)
-        if parsed_timestamp is None:
-            continue
-        points.append((parsed_timestamp, float(value)))
-
-    if not points:
-        return {}
-
-    points.sort(key=lambda item: item[0])
-    deltas = [
-        int((points[index + 1][0] - points[index][0]).total_seconds())
-        for index in range(len(points) - 1)
-        if 0 < int((points[index + 1][0] - points[index][0]).total_seconds()) <= 21600
-    ]
-    fallback_step_seconds = min(deltas) if deltas else 3600
-
-    daily_wh: Dict[str, float] = {}
-    for index, (timestamp, power_w) in enumerate(points):
-        if index + 1 < len(points):
-            step_seconds = int((points[index + 1][0] - timestamp).total_seconds())
-            if step_seconds <= 0 or step_seconds > 21600:
-                step_seconds = fallback_step_seconds
-        else:
-            step_seconds = fallback_step_seconds
-        date_key = timestamp.date().isoformat()
-        daily_wh[date_key] = daily_wh.get(date_key, 0.0) + max(0.0, power_w) * (step_seconds / 3600.0)
-
-    return {date_key: energy_wh / 1000.0 for date_key, energy_wh in daily_wh.items()}
-
-
 def _filter_values_for_date(values: Dict[str, float], target_date) -> Dict[str, float]:
     day_values: Dict[str, float] = {}
     for timestamp, value in values.items():
@@ -75,8 +41,7 @@ class ForecastValueStore(ValueStore[ForecastState]):
 
     def update(self):
         values = self.state.forecast_values or {}
-        provider_daily_kwh = self.state.daily_kwh or {}
-        daily_kwh = provider_daily_kwh if provider_daily_kwh else _calculate_daily_kwh(values)
+        daily_kwh = self.state.daily_kwh or {}
         today_date = datetime.now().date()
         tomorrow_date = datetime.now().date() + timedelta(days=1)
         today_values = _filter_values_for_date(values, today_date)
